pyproject.toml
[project]
name = "cambot"
version = "0.1.0"
description = "Security camera monitoring agent powered by Claude"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.42.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "python-telegram-bot>=21.0",
]

[project.scripts]
cambot = "cambot.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
